+++ packages/@webex/contact-center/test/unit/spec/services/EntryPointAPI.ts
@@ -1,27 +1,27 @@
-import EntryPointAPI from '../../../src/services/EntryPointAPI';
-import {HTTP_METHODS, WebexSDK, IHttpResponse} from '../../../src/types';
-import {METRIC_EVENT_NAMES} from '../../../src/metrics/constants';
-import WebexRequest from '../../../src/services/core/WebexRequest';
-import MetricsManager from '../../../src/metrics/MetricsManager';
-import LoggerProxy from '../../../src/logger-proxy';
+import EntryPointAPI from '../../../../src/services/EntryPointAPI';
+import {HTTP_METHODS, WebexSDK, IHttpResponse} from '../../../../src/types';
+import {METRIC_EVENT_NAMES} from '../../../../src/metrics/constants';
+import WebexRequest from '../../../../src/services/core/WebexRequest';
+import MetricsManager from '../../../../src/metrics/MetricsManager';
+import LoggerProxy from '../../../../src/logger-proxy';
 
-jest.mock('../../../src/services/core/WebexRequest');
-jest.mock('../../../src/metrics/MetricsManager');
-jest.mock('../../../src/logger-proxy');
+jest.mock('../../../../src/metrics/MetricsManager');
+jest.mock('../../../../src/logger-proxy');
 
 describe('EntryPointAPI', () => {
   let entryPointAPI: EntryPointAPI;
   let mockWebex: WebexSDK;
-  let mockWebexRequest: jest.Mocked<WebexRequest>;
   let mockMetricsManager: jest.Mocked<MetricsManager>;
 
   beforeEach(() => {
     jest.clearAllMocks();
 
+    (WebexRequest as any).instance = undefined;
     mockWebex = {
       credentials: {
         getOrgId: jest.fn().mockReturnValue('test-org-id'),
       },
+      request: jest.fn(),
       internal: {
         newMetrics: {
           submitBehavioralEvent: jest.fn(),
@@ -32,11 +32,6 @@ describe('EntryPointAPI', () => {
       ready: true,
       once: jest.fn(),
     } as unknown as WebexSDK;
-
-    mockWebexRequest = {
-      request: jest.fn(),
-    } as unknown as jest.Mocked<WebexRequest>;
-    (WebexRequest.getInstance as jest.Mock).mockReturnValue(mockWebexRequest);
 
     mockMetricsManager = {
       trackEvent: jest.fn(),
@@ -49,7 +44,7 @@ describe('EntryPointAPI', () => {
 
   describe('constructor', () => {
     it('should initialize with all required dependencies', () => {
-      expect(WebexRequest.getInstance).toHaveBeenCalledWith({webex: mockWebex});
+      expect(WebexRequest.getInstance({webex: mockWebex})).toBeDefined();
       expect(MetricsManager.getInstance).toHaveBeenCalledWith({webex: mockWebex});
     });
   });
@@ -98,11 +93,11 @@ describe('EntryPointAPI', () => {
     });
 
     it('should fetch entry points successfully with default parameters', async () => {
-      mockWebexRequest.request.mockResolvedValue(mockResponse);
+      (mockWebex.request as jest.Mock).mockResolvedValue(mockResponse);
 
       const result = await entryPointAPI.getEntryPoints();
 
-      expect(mockWebexRequest.request).toHaveBeenCalledWith({
+      expect(mockWebex.request).toHaveBeenCalledWith({
         service: 'wcc-api-gateway',
         resource: '/organization/test-org-id/v2/entry-point?page=0&pageSize=100&sortOrder=asc',
         method: HTTP_METHODS.GET,
@@ -122,12 +117,18 @@ describe('EntryPointAPI', () => {
         },
         ['behavioral']
       );
-      expect(LoggerProxy.info).toHaveBeenCalled();
-      expect(LoggerProxy.log).toHaveBeenCalled();
+      expect(LoggerProxy.info).toHaveBeenCalledWith(
+        'Fetching entry points - orgId: test-org-id, page: 0, pageSize: 100, isSearchRequest: false',
+        {module: 'EntryPointAPI', method: 'getEntryPoints'}
+      );
+      expect(LoggerProxy.log).toHaveBeenCalledWith(
+        `Making API request to fetch entry points - resource: /organization/test-org-id/v2/entry-point?page=0&pageSize=100&sortOrder=asc, service: wcc-api-gateway`,
+        {module: 'EntryPointAPI', method: 'getEntryPoints'}
+      );
     });
 
     it('should fetch entry points with custom parameters', async () => {
-      mockWebexRequest.request.mockResolvedValue(mockResponse);
+      (mockWebex.request as jest.Mock).mockResolvedValue(mockResponse);
 
       const params = {
         page: 1,
@@ -141,7 +142,7 @@ describe('EntryPointAPI', () => {
 
       await entryPointAPI.getEntryPoints(params);
 
-      expect(mockWebexRequest.request).toHaveBeenCalledWith({
+      expect(mockWebex.request).toHaveBeenCalledWith({
         service: 'wcc-api-gateway',
         resource: '/organization/test-org-id/v2/entry-point?page=1&pageSize=25&sortOrder=desc&search=test&filter=type%3D%3D%22voice%22&attributes=id%2Cname&sortBy=name',
         method: HTTP_METHODS.GET,
@@ -162,7 +163,7 @@ describe('EntryPointAPI', () => {
     });
 
     it('should handle API errors and track metrics', async () => {
-      mockWebexRequest.request.mockRejectedValue(new Error('Internal Server Error'));
+      (mockWebex.request as jest.Mock).mockRejectedValue(new Error('Internal Server Error'));
 
       await expect(entryPointAPI.getEntryPoints()).rejects.toThrow('Internal Server Error');
 
@@ -180,56 +181,22 @@ describe('EntryPointAPI', () => {
       expect(LoggerProxy.error).toHaveBeenCalled();
     });
 
-    it('should handle network errors and track metrics', async () => {
-      const networkError = new Error('Network error');
-      mockWebexRequest.request.mockRejectedValue(networkError);
 
-      await expect(entryPointAPI.getEntryPoints()).rejects.toThrow('Network error');
-
-      expect(mockMetricsManager.trackEvent).toHaveBeenCalledWith(
-        METRIC_EVENT_NAMES.ENTRYPOINT_FETCH_FAILED,
-        {
-          orgId: 'test-org-id',
-          error: 'Network error',
-          isSearchRequest: false,
-          page: 0,
-          pageSize: 100,
-        },
-        ['behavioral']
-      );
-      expect(LoggerProxy.error).toHaveBeenCalledWith(
-        'Failed to fetch entry points',
-        {
-          module: 'EntryPointAPI',
-          method: 'getEntryPoints',
-          data: {
-            orgId: 'test-org-id',
-            error: 'Network error',
-            isSearchRequest: false,
-            page: 0,
-            pageSize: 100,
-          },
-          error: networkError,
-        }
-      );
-    });
 
     it('should not track metrics for subsequent pages in simple pagination', async () => {
-      mockWebexRequest.request.mockResolvedValue(mockResponse);
+      (mockWebex.request as jest.Mock).mockResolvedValue(mockResponse);
 
-      await entryPointAPI.getEntryPoints({page: 2});
+      const result = await entryPointAPI.getEntryPoints({page: 2});
+      expect(result).toEqual(mockResponse.body);
 
-      expect(mockMetricsManager.trackEvent).not.toHaveBeenCalledWith(
-        METRIC_EVENT_NAMES.ENTRYPOINT_FETCH_SUCCESS,
-        expect.any(Object),
-        expect.any(Array)
-      );
+      expect(mockMetricsManager.trackEvent).not.toHaveBeenCalled();
     });
 
     it('should track metrics for search requests on any page', async () => {
-      mockWebexRequest.request.mockResolvedValue(mockResponse);
+      (mockWebex.request as jest.Mock).mockResolvedValue(mockResponse);
 
-      await entryPointAPI.getEntryPoints({page: 2, search: 'test'});
+      const result2 = await entryPointAPI.getEntryPoints({page: 2, search: 'test'});
+      expect(result2).toEqual(mockResponse.body);
 
       expect(mockMetricsManager.trackEvent).toHaveBeenCalledWith(
         METRIC_EVENT_NAMES.ENTRYPOINT_FETCH_SUCCESS,
